requirements/local.txt
# Local development dependencies go here
-r base.txt
coverage==3.7.1
django-discover-runner==0.4
django-debug-toolbar==1.0.1
Sphinx==1.2.2
fabric==1.8.3
